POO get et set

+++ delete.php
@@ -2,25 +2,46 @@
 
 <?php
 
-// Inclusion de mon Header
+// Inclusion de mon Header et appelle de ma SESSION_START
 $pageTitle = 'Page de connexion';
 require 'header.php';
 
-session_start();
 require 'db.php';
 
-// Protection
+// Protection : utilisateur connecté
 if (!isset($_SESSION['user_id'])) {
     header('Location: login.php');
     exit;
 }
 
 // Vérifier l'ID
-if (!isset($_GET['id'])) {
-    die("ID manquant");
+
+// EXO PARTIE 1
+// if (!isset($_GET['id'])) {
+//     die("ID manquant");
+// }
+
+// $id = (int) $_GET['id'];
+
+// EXO PARTIE 2
+if (!isset($_GET['id']) || empty($_GET['id'])) {
+    header('Location: list_users.php?msg=invalid_id');
+    exit;
 }
 
-$id = (int) $_GET['id'];
+// Sécuriser l'ID - Validation
+$id = filter_var($_GET['id'], FILTER_VALIDATE_INT);
+
+if ($id === false) {
+    header('Location: list_users.php?msg=invalid_id');
+    exit;
+}
+
+// Empêcher la suppression de soi-même
+if ($id === (int) $_SESSION['user_id']) {
+    header('Location: list_users.php?msg=self_delete_forbidden');
+    exit;
+}
 
 // Suppression
 $stmt = $pdo->prepare("DELETE FROM users WHERE id = ?");
